# ops-timer-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 运维计时管理系统 API
  description: |
    面向运维场景的计时、项目、待办与通知管理 REST API。

    ## 认证
    - **JWT**：在请求头添加 `Authorization: Bearer <token>`，`token` 由 `POST /api/v1/auth/login` 或 OAuth 回调获得。
    - **API Token**：在请求头添加 `X-API-Token: <api_token>`，适用于脚本与第三方集成（见 `GET /api/v1/auth/token`）。

    ## 统一响应体
    除 `204 No Content`、OAuth 重定向及少数裸 JSON 外，成功响应体为：
    ```json
    { "code": 0, "message": "success", "data": { ... } }
    ```
    分页列表额外包含 `meta`：`page`、`page_size`、`total`、`total_pages`。

    ## 业务错误码（`code` 字段）
    | code | 说明 |
    |------|------|
    | 0 | 成功 |
    | 40001 | 参数校验失败 |
    | 40101 | 认证缺失 |
    | 40102 | Token 无效 |
    | 40103 | 凭据错误 |
    | 40104 | 登录锁定（限流） |
    | 40401 | 资源不存在 |
    | 42201 | 业务规则错误 |
    | 50001 | 服务器内部错误 |
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: "{baseUrl}"
    variables:
      baseUrl:
        default: http://localhost:8080
        description: 服务根地址（不含尾部斜杠）
...
paths:
...
  /api/v1/auth/oauth/login:
    get:
      tags: [OAuth]
      summary: 跳转至 IdP 授权页
      description: |
        未启用 OAuth 时返回 `503`。成功时返回 **302**，`Location` 为 IdP 授权 URL。
      operationId: oauthLogin
      responses:
        "302":
          description: 重定向到 OAuth 提供商
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          description: OAuth 未启用
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: OAuth 未启用
  /api/v1/auth/oauth/callback:
    get:
      tags: [OAuth]
      summary: OAuth 授权回调（由 IdP 重定向调用）
      description: |
        成功时 **302** 重定向到前端 `frontend_url/oauth/callback?token=...`。
        失败时可能 **302** 带 `error` 查询参数，或在未配置前端 URL 时返回 **400** JSON。
      operationId: oauthCallback
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
        - name: error_description
          in: query
          schema:
            type: string
      responses:
        "302":
          description: 重定向到前端或错误页
        "400":
          description: 参数错误或无法重定向（无 frontend_url）
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 使用 `POST /api/v1/auth/login` 或 OAuth 回调获得的 access token。
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Token
      description: 长期 API Token，可在「设置」中查看与轮换。
...
  responses:
...
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
  schemas:
...
    EnvelopeError:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...